Feat(test-common): Move basic channel tests into channel sub-project
