+++ test-common/build.gradle
@@ -13,6 +13,8 @@ dependencies {
     implementation platform(project(':alert-platform'))
     implementation project(':alert-common')
 
+    implementation 'com.synopsys.integration:integration-common'
+
     implementation 'org.springframework:spring-test'
     implementation 'org.springframework.security:spring-security-test'
     implementation 'org.springframework.boot:spring-boot-test-autoconfigure'
